refactor(library/ContentDescriptor): renames directory from "/MediaType" to "/ContentDescriptor"
