+++ src/library/ContentDescriptor/StructuredSyntaxNameSuffix/Any.ts
