+++ src/library/ContentDescriptor/StructuredSyntaxNameSuffix/all.ts
