+++ src/library/ContentDescriptor/StructuredSyntaxNameSuffix/exports.ts
