+++ src/library/ContentDescriptor/StructuredSyntaxNameSuffix/index.ts
