+++ src/library/ContentDescriptor/TopLevel/Discrete/Any.ts
