+++ src/library/ContentDescriptor/index.ts
